Moved notification utils to other package
